refactor: mv code used by new endpoint out from email-submission
- validateAttachments + ParsedResponsesObject
